requirements.txt
blessed==1.15.0
botocore==1.12.71
cached-property==1.4.3
cachetools==3.0.0
cement==2.8.2
certifi==2018.11.29
chardet==3.0.4
colorama==0.3.9
concurrent-log-handler==0.9.12
ConcurrentLogHandler==0.9.1
defusedxml==0.5.0
Django==2.0.7
django-braces==1.13.0
django-oauth-toolkit==1.2.0
django-templated-mail==1.1.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djoser==1.2.0
docker==3.4.1
docker-compose==1.21.2
docker-pycreds==0.3.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
gcloud==0.17.0
google-api-core==1.7.0
google-auth==1.6.2
google-cloud-core==0.29.1
google-cloud-logging==1.9.1
googleapis-common-protos==1.5.6
grpcio==1.17.1
httplib2==0.11.3
idna==2.6
jmespath==0.9.3
jsonschema==2.6.0
jws==0.1.3
mysqlclient==1.3.13
oauth2client==3.0.0
oauthlib==2.1.0
pathspec==0.5.5
protobuf==3.6.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycryptodome==3.4.3
PyJWT==1.6.4
Pyrebase==3.0.27
python-dateutil==2.7.3
python-decouple==3.1
python-google-places==1.4.1
python-jwt==2.0.1
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.5
PyYAML==3.13
requests==2.21.0
requests-oauthlib==1.0.0
requests-toolbelt==0.7.0
rest-social-auth==1.3.1
rsa==3.4.2
semantic-version==2.5.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
tabulate==0.7.5
termcolor==1.1.0
texttable==0.9.1
urllib3==1.22
wcwidth==0.1.7
websocket-client==0.48.0
whitenoise==4.1.2
